refactory anomalies form and redux
