+++ package.json
@@ -26,6 +26,7 @@
     "react-dom": "17.0.2",
     "react-html5-camera-photo": "1.5.5",
     "react-moment": "1.1.1",
+    "react-redux": "7.2.7",
     "react-router-dom": "6.2.1",
     "react-scripts": "4.0.0",
     "redux-logger": "3.0.6",
