+++ src/App.js
@@ -3,18 +3,20 @@ import "./styles.css";
 import { useContext } from "react";
 import SignIn from "./features/Login";
 import Navbar from "./components/Navbar";
-import { CredentialContext } from "./contex/StoreContext";
+
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Container from "@mui/material/Container";
 import Home from "./features/Home";
 import Anomalies from "./features/Anomalies";
 import NewUser from "./features/NewUser";
 import Update from "./features/UpDate";
+import { useSelector } from "react-redux";
 
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 
 export default function App() {
-  const { user } = useContext(CredentialContext);
+  const user = useSelector((state) => state.user);
+
   const theme = createTheme({
     palette: {
       primary: {
@@ -36,18 +38,21 @@ export default function App() {
           }}
         >
           <Router>
-            {user && <Navbar />}
+            {user.name && <Navbar />}
 
             <Routes>
               <Route path="/" element={<SignIn />} />
               <Route path="anomalies" element={<Anomalies />} />
-              <Route path="newuser" element={user ? <NewUser /> : <SignIn />} />
+              <Route
+                path="newuser"
+                element={user.name ? <NewUser /> : <SignIn />}
+              />
               <Route path="update" element={<Update />} />
-              <Route path="home" element={user ? <Home /> : <SignIn />} />
-              {/*  <Route
-              path="anomalies"
-              element={user ? <Anomalies /> : <SignIn />}
-            /> */}
+              <Route path="home" element={user.name ? <Home /> : <SignIn />} />
+              <Route
+                path="anomalies"
+                element={user ? <Anomalies /> : <SignIn />}
+              />
             </Routes>
           </Router>
         </Container>
