+++ src/features/UpDate.js
@@ -1,53 +1,40 @@
 import { useLocation } from "react-router-dom";
-import UpdateFrom from "../components/UpdateFrom";
+import AnomaliesForm from "../components/AnomaliesForm";
 import { Typography } from "@mui/material";
-import { db } from "../config/firebase_config";
-import { doc, setDoc, updateDoc, increment, getDoc } from "firebase/firestore";
-import { useEffect, useState, useContext } from "react";
-import { CredentialContext } from "../contex/StoreContext";
+import { useSelector, useDispatch } from "react-redux";
+import { updateItem } from "../slice/itemSlice";
+import { useEffect, useState } from "react";
+
 import moment from "moment";
 
 function UpDate() {
   const { state } = useLocation();
-  const { user, lists } = useContext(CredentialContext);
+  const user = useSelector((state) => state.user);
+  const items = useSelector((state) => state.items);
+  const dispatch = useDispatch();
 
   console.log("state", state);
   const [item, setItem] = useState();
 
-  const getItemFromLists = () => {
-    const finder = lists.find((e) => e.id === state.id);
+  const getItemFromLists = async () => {
+    const finder = await items.find((e) => e.id === state.id);
+
     setItem(finder);
   };
 
-  const getItem = async (id) => {
-    const docRef = doc(db, "anomalies", id);
-    const docSnap = await getDoc(docRef);
-    if (docSnap.exists()) {
-      console.log("Document data:", docSnap.data());
-      setItem(docSnap.data());
-    } else {
-      // doc.data() will be undefined in this case
-      console.log("No such document!");
-    }
-  };
-
   const handleSubmit = async () => {
-    console.log("item anomalies", item);
-    if (item.category === "" || item.priority === "") {
-      return alert("devi selezionare una categoria");
-    }
-    const itemRef = doc(db, "anomalies", `${item.item_ref}`);
-    const res = await setDoc(itemRef, {
+    setItem({
       ...item,
       lastUpdate: moment().format("DD/MM/YYYY"),
       updateBy: user.name
     });
-    console.log("Document written with ID: ", itemRef.id);
-    return res;
+
+    dispatch(updateItem({ item }));
   };
 
   useEffect(() => {
     getItemFromLists();
+    console.log("update", item);
     /* getItem(state.id); */
     return () => {};
   }, []);
@@ -55,13 +42,15 @@ function UpDate() {
   return (
     <div>
       <Typography variant="h4">Update item {state.id}</Typography>
-
-      <UpdateFrom
-        item={item}
-        setItem={setItem}
-        user={user}
-        handleSubmit={handleSubmit}
-      />
+      {item && (
+        <AnomaliesForm
+          update={true}
+          item={item}
+          setItem={setItem}
+          user={user}
+          handleSubmit={handleSubmit}
+        />
+      )}
     </div>
   );
 }
